requirements.txt
spacy==3.3.0
spacy-transformers==1.1.5
numba==0.55.1
torchtext==0.13.0
pytorch-lightning==1.6.4
spacy-conll==3.2.0
scikit-learn==1.1.1
#  h5
h5py==3.6.0
# local
pylint==2.13.8
black==22.3.0
# Notebook
ipywidgets==7.7.1
tqdm==4.64.0
tensorboard==2.9.1
# Test
pytest==7.1.2